acrn-hypervisor: launch_uos.sh: fix a typo
The /dev/vbs_ipu is used by the backend driver of IPU mediator on SOS
kernel.

If the file of /dev/vbs_ipu exists, it indicates that IPU works in
mediator mode. Other it will fall back to the pass-through mode. But it
should be -e instead of -f check the file of /dev/vbs_ipu correctly.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>